Added YouGroupSettings (PoomSmart)
Whenever I updated the Submodules, I saw this when I was looking through the changes. And the weirder thing is that my settings changed and I’m certain and hope that this fixes the problems I’ve been having in the Settings. If it doesn’t then I don’t know.